Déplace `/api/token` dans les routes explicitement publiques
