+++ src/routes/publiques/routesApiPublique.js
@@ -6,8 +6,45 @@ const routesApiPublique = ({
   referentiel,
   depotDonnees,
   adaptateurAnnuaire,
+  adaptateurTracking,
 }) => {
   const routes = express.Router();
+
+  routes.post('/token', (requete, reponse, suite) => {
+    const login = requete.body.login?.toLowerCase();
+    const { motDePasse } = requete.body;
+    depotDonnees
+      .utilisateurAuthentifie(login, motDePasse)
+      .then((utilisateur) => {
+        if (utilisateur) {
+          const token = utilisateur.genereToken();
+          requete.session.token = token;
+          depotDonnees.homologations(utilisateur.id).then((services) =>
+            adaptateurTracking.envoieTrackingConnexion(utilisateur.email, {
+              nombreServices: services.length,
+            })
+          );
+          depotDonnees
+            .lisParcoursUtilisateur(utilisateur.id)
+            .then((parcoursUtilisateur) => {
+              const nouvelleFonctionnalite =
+                parcoursUtilisateur.recupereNouvelleFonctionnalite();
+              parcoursUtilisateur.enregistreDerniereConnexionMaintenant();
+              depotDonnees
+                .sauvegardeParcoursUtilisateur(parcoursUtilisateur)
+                .then(() =>
+                  reponse.json({
+                    utilisateur: utilisateur.toJSON(),
+                    nouvelleFonctionnalite,
+                  })
+                );
+            });
+        } else {
+          reponse.status(401).send("L'authentification a échoué");
+        }
+      })
+      .catch(suite);
+  });
 
   routes.get('/dureeSession', (_requete, reponse) => {
     reponse.send({ dureeSession: DUREE_SESSION });
